+++ tests/weather_status_test.py
@@ -14,10 +14,10 @@ def test_available(nc):
 def test_get_set_location(nc):
     nc.weather_status.set_location(longitude=0.0, latitude=0.0)
     loc = nc.weather_status.get_location()
-    assert loc["latitude"] == 0.0
-    assert loc["longitude"] == 0.0
-    assert isinstance(loc["address"], str)
-    assert isinstance(loc["mode"], int)
+    assert loc.latitude == 0.0
+    assert loc.longitude == 0.0
+    assert isinstance(loc.address, str)
+    assert isinstance(loc.mode, int)
     try:
         assert nc.weather_status.set_location(address="Paris, 75007, France")
     except NextcloudException as e:
@@ -25,25 +25,25 @@ def test_get_set_location(nc):
             pytest.skip("Some network problem on the host")
         raise e from None
     loc = nc.weather_status.get_location()
-    assert loc["latitude"]
-    assert loc["longitude"]
-    if loc["address"].find("Unknown") != -1:
+    assert loc.latitude
+    assert loc.longitude
+    if loc.address.find("Unknown") != -1:
         pytest.skip("Some network problem on the host")
-    assert loc["address"].find("Paris") != -1
+    assert loc.address.find("Paris") != -1
     assert nc.weather_status.set_location(latitude=41.896655, longitude=12.488776)
     loc = nc.weather_status.get_location()
-    assert loc["latitude"] == 41.896655
-    assert loc["longitude"] == 12.488776
-    if loc["address"].find("Unknown") != -1:
+    assert loc.latitude == 41.896655
+    assert loc.longitude == 12.488776
+    if loc.address.find("Unknown") != -1:
         pytest.skip("Some network problem on the host")
-    assert loc["address"].find("Rom") != -1
+    assert loc.address.find("Rom") != -1
     assert nc.weather_status.set_location(latitude=41.896655, longitude=12.488776, address="Paris, France")
     loc = nc.weather_status.get_location()
-    assert loc["latitude"] == 41.896655
-    assert loc["longitude"] == 12.488776
-    if loc["address"].find("Unknown") != -1:
+    assert loc.latitude == 41.896655
+    assert loc.longitude == 12.488776
+    if loc.address.find("Unknown") != -1:
         pytest.skip("Some network problem on the host")
-    assert loc["address"].find("Rom") != -1
+    assert loc.address.find("Rom") != -1
 
 
 @pytest.mark.parametrize("nc", NC_TO_TEST)
@@ -55,7 +55,7 @@ def test_get_set_location_no_lat_lon_address(nc):
 @pytest.mark.parametrize("nc", NC_TO_TEST)
 def test_get_forecast(nc):
     nc.weather_status.set_location(latitude=41.896655, longitude=12.488776)
-    if nc.weather_status.get_location()["address"].find("Unknown") != -1:
+    if nc.weather_status.get_location().address.find("Unknown") != -1:
         pytest.skip("Some network problem on the host")
     forecast = nc.weather_status.get_forecast()
     assert isinstance(forecast, list)
@@ -78,9 +78,9 @@ def test_get_set_favorites(nc):
 @pytest.mark.parametrize("nc", NC_TO_TEST)
 def test_set_mode(nc):
     nc.weather_status.set_mode(WeatherLocationMode.MODE_BROWSER_LOCATION)
-    assert nc.weather_status.get_location()["mode"] == WeatherLocationMode.MODE_BROWSER_LOCATION.value
+    assert nc.weather_status.get_location().mode == WeatherLocationMode.MODE_BROWSER_LOCATION.value
     nc.weather_status.set_mode(WeatherLocationMode.MODE_MANUAL_LOCATION)
-    assert nc.weather_status.get_location()["mode"] == WeatherLocationMode.MODE_MANUAL_LOCATION.value
+    assert nc.weather_status.get_location().mode == WeatherLocationMode.MODE_MANUAL_LOCATION.value
 
 
 @pytest.mark.parametrize("nc", NC_TO_TEST)